Move interfaces related to serializer context in another file
